[DEV-5536] moving file to covid dir
